Modularize more conv ops.
