+++ tfjs-backend-cpu/src/register_all_kernels.ts
@@ -38,6 +38,11 @@ import {clipConfig} from './kernels/Clip';
 import {complexConfig} from './kernels/Complex';
 import {concatConfig} from './kernels/Concat';
 import {conv2DConfig} from './kernels/Conv2D';
+import {conv2DBackpropFilterConfig} from './kernels/Conv2DBackpropFilter';
+import {conv2DBackpropInputConfig} from './kernels/Conv2DBackpropInput';
+import {conv3DConfig} from './kernels/Conv3D';
+import {conv3DBackpropFilterV2Config} from './kernels/Conv3DBackpropFilterV2';
+import {conv3DBackpropInputV2Config} from './kernels/Conv3DBackpropInputV2';
 import {cosConfig} from './kernels/Cos';
 import {coshConfig} from './kernels/Cosh';
 import {dilation2dConfig} from './kernels/Dilation2D';
@@ -119,7 +124,12 @@ const kernelConfigs: KernelConfig[] = [
   clipConfig,
   complexConfig,
   concatConfig,
+  conv2DBackpropInputConfig,
+  conv2DBackpropFilterConfig,
   conv2DConfig,
+  conv3DBackpropInputV2Config,
+  conv3DBackpropFilterV2Config,
+  conv3DConfig,
   cosConfig,
   coshConfig,
   dilation2dConfig,
